LayoutTests/media/modern-media-controls/tracks-support/tracks-support-show-and-populate-panel-expected.txt

Showing and populating the tracks panel.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Audio
  English Sound
  French Sound
  Spanish Sound

Subtitles
  Off
  Auto (Recommended)
  English Subtitles
  French Subtitles
  German Subtitles
  Spanish Subtitles

PASS successfullyParsed is true

TEST COMPLETE
